refactor(tx): rename AddInputs to AddConfigGenerator and support burning

+++ src/Chrysalis.Tx/Models/Options.cs
@@ -88,7 +88,7 @@ public record OutputOptions
 public record MintOptions<T>
 {
     public string Policy { get; set; } = string.Empty;
-    public Dictionary<string, ulong> Assets { get; set; } = [];
+    public Dictionary<string, int> Assets { get; set; } = [];
     public RedeemerMap? Redeemer { get; set; }
     public Func<InputOutputMapping, T, Redeemer<CborBase>>? RedeemerBuilder { get; set; }
     public MintOptions<T> SetRedeemerBuilder<TData>(RedeemerDataBuilder<T, TData> factory)
